// types/address/addressType.ts
export interface IAddressData {
  /** 시도 */
  sido: string;
  /** 시군구 */
  sigungu: string;
  /** 법정동/법정리 */
  bname: string;
  /** 법정리 읍/면 /*/
  bname1: string;
  /** 도로명주소 */
  roadAddress: string;
  /** 건물이름 */
  buildingName: string;
  /** 우편번호 */
  zonecode: string;
}